requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.10.0
pydantic-settings>=2.4.0
google-cloud-firestore>=2.16.0
google-cloud-aiplatform>=1.71.1
vertexai>=1.71.1
python-dotenv>=1.0.1
httpx>=0.27.0
pytest>=8.2.0
pytest-asyncio>=0.23.6
cachetools>=5.3.3
